feat:FastApi 통신, kis 폴백로직 구현 1. 폴백: DB조회 -> 없으면 외부api kis먼저 -> 실패시 marketstack -> 성공시 db insert -> 실패시 error 던짐 2. DTO fastapi, spring간 통일 3. postman, curl 테스트, 텍스트 출력 성공

+++ backend/src/main/java/com/qaima/dto/MarketStackTickersResponse.java
@@ -2,6 +2,7 @@ package com.qaima.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -25,8 +26,8 @@ public class MarketStackTickersResponse {
         private Boolean has_intraday;       // intraday 데이터 보유 여부
         private Boolean has_eod;            // EOD 데이터 보유 여부
         private StockExchange stock_exchange;
-        private Double price;               // 현재가
-        private Double changeRate;          // 등락률 (%)
+        private BigDecimal price;               // 현재가
+        private BigDecimal changeRate;          // 등락률 (%)
     }
 
     @Data
